+++ packages/synapse-interface/components/StateManagedBridge/InputContainer.tsx
@@ -30,6 +30,7 @@ import { useGasEstimator } from '../../utils/hooks/useGasEstimator'
 import { getParsedBalance } from '@/utils/getParsedBalance'
 import { MaxButton } from './MaxButton'
 import { formatAmount } from '../../utils/formatAmount'
+import { useWalletState } from '@/slices/wallet/hooks'
 
 export const inputRef = React.createRef<HTMLInputElement>()
 
@@ -37,14 +38,9 @@ export const InputContainer = () => {
   const dispatch = useAppDispatch()
   const { chain, isConnected } = useAccount()
   const { balances } = usePortfolioState()
-  const {
-    fromChainId,
-    toChainId,
-    fromToken,
-    toToken,
-    fromValue,
-    isWalletPending,
-  } = useBridgeState()
+  const { fromChainId, toChainId, fromToken, toToken, fromValue } =
+    useBridgeState()
+  const { isWalletPending } = useWalletState()
   const [showValue, setShowValue] = useState('')
   const [hasMounted, setHasMounted] = useState(false)
 
@@ -174,7 +170,8 @@ export const InputContainer = () => {
               !isConnected ||
               !hasValidInputSelections ||
               isLoading ||
-              isInputMax
+              isInputMax ||
+              isWalletPending
             }
           />
         </div>
@@ -185,6 +182,7 @@ export const InputContainer = () => {
 
 const FromChainSelector = () => {
   const { fromChainId } = useBridgeState()
+  const { isWalletPending } = useWalletState()
 
   return (
     <ChainSelector
@@ -195,12 +193,14 @@ const FromChainSelector = () => {
       itemListFunction={useFromChainListArray}
       setFunction={setFromChainId}
       action="Bridge"
+      disabled={isWalletPending}
     />
   )
 }
 
 const FromTokenSelector = () => {
   const { fromToken } = useBridgeState()
+  const { isWalletPending } = useWalletState()
 
   return (
     <TokenSelector
@@ -211,6 +211,7 @@ const FromTokenSelector = () => {
       itemListFunction={useFromTokenListArray}
       setFunction={setFromToken}
       action="Bridge"
+      disabled={isWalletPending}
     />
   )
 }