+++ packages/synapse-interface/store/reducer.ts
@@ -16,6 +16,7 @@ import poolUserData from '@/slices/poolUserDataSlice'
 import poolWithdraw from '@/slices/poolWithdrawSlice'
 import priceData from '@/slices/priceDataSlice'
 import gasData from '@/slices/gasDataSlice'
+import wallet from '@/slices/wallet/reducer'
 import { api } from '@/slices/api/slice'
 import { RootActions } from '@/slices/application/actions'
 
@@ -46,6 +47,7 @@ export const appReducer = combineReducers({
   poolWithdraw,
   priceData,
   gasData,
+  wallet,
   [api.reducerPath]: api.reducer,
   ...persistedReducers,
 })